[UPDATE] local stored user data in FE and tickets page for clients

+++ api/Models/SlotMachine/SlotMachineSpinReq.cs
@@ -8,7 +8,7 @@ namespace WebApi.Models
         public int UserId { get; set; }
 
         [Required]
-        public int Bet { get; set; }
+        public float Bet { get; set; }
 
         [Required]
         public string Game { get; set; }
